requirements.txt
newsapi-python
openai
python-telegram-bot
